Updates project structure, by moving up the preconf package.
